Extract FixerWorkerRunner from FixerWorkerCommand
Moves everything FixerWorkerCommand does after CommandHelper::begin() —
connecting back to FixerApplication, restoring the result cache,
running the analyser and streaming results — into a reusable
FixerWorkerRunner service. FixerWorkerCommand now only boots and
delegates, making the post-boot logic callable without a re-boot.
